Q&A_Chatbot_Conversational_Memory/requirements.txt
langchain-openai
pandas
numpy
python-dotenv
langchain-community
langchain
langchain_chroma
langchain-core
streamlit
pypdf
langchain_groq
langchain_huggingface
sentence-transformers